refactor: 💡 contractsView level 2 is done
