+++ packages/frontend/src/App.jsx
@@ -25,8 +25,8 @@ import AuthorsUsersScreen from './components/ui/admin/AuthorsUsersScreen/Authors
 import FooterMember from './utils/components/FooterMember';
 import MembersMain from './components/Members/MembersMain';
 import FullPublication from './components/Members/ShowPublication/FullPublication';
-import AllContracts from './components/SalesScreens/ContractViews/AllContracts/AllContract';
-import AllInvoices from './components/SalesScreens/ContractViews/AllInvoices/AllInvoices';
+import AllContracts from './components/ui/salesman/contractViews/allContracts/AllContract/AllContract';
+import AllInvoices from './components/ui/salesman/contractViews/allInvoices/AllInvoices/AllInvoices';
 import AllNotifications from './components/Members/Notifications/AllNotifications';
 import AccountSettings from './components/AccountSettings/AccountSettings';
