+++ packages/frontend/src/components/ui/salesman/contractViews/allInvoices/AllInvoices/AllInvoices.style.js
@@ -1,0 +1,5 @@
+import { makeStyles } from '@material-ui/core/styles';
+
+const useStyles = makeStyles(() => ({}));
+
+export default useStyles;
